Phase 1 Step 4: extract PricesService + analytics/indicators
Move the 12 price/technical MCP tools and the REST /ohlcv, /technicals,
/signals/technical, /securities/screen routes into PricesService
(constructor-injected with OhlcvRepository, YFinanceGateway, OptionsStore,
SentimentStore). Bodies copied verbatim — behavioral parity is the contract.

- quantcore/analytics/indicators.py: pure safe_float/rsi_series/macd_series,
  the single home for the REST-table RSI/MACD math (deletes api/app.py's
  third copy). MCP get_rsi/get_macd keep their adjust=False bodies verbatim
  inside the service — intentionally not unified.
- OhlcvRepository.daily_bars_for_symbols(): the screener's one-shot SQL.
- registry: hoist OptionsStore() to a shared local; wire prices=PricesService.
- stock_price_server.py: 12 tools become thin adapters (docstrings verbatim);
  _safe_int/_options_store/_chain_side_full kept for the options tools (Step 5).
- api/app.py: 4 routes repointed; local _safe_float/_compute_rsi/_compute_macd
  deleted.

Verified vs QUANTCORE_TEST_DB_DSN: parity 19/19 (sole diff is non-deterministic
set-iteration order in a ValueError string, identical content); 30 tests green;
5 MCP servers boot (stock-price 23 tools); Flask 37 routes.
