+++ docs/proposals/phase1-migration-plan.md
@@ -12,7 +12,7 @@ Update this table as part of every step-commit. Any dev machine can resume by re
 | 1 | Microstructure service | DONE | — | 2026-06-12 | Template step. quantcore/gateways/yfinance_gateway.py (YFinanceGateway: ticker_info w/ 15s timeout, fast_info, expirations, option_chain) + quantcore/services/microstructure.py (MicrostructureService: get_short_interest/get_dark_pool/get_bid_ask_spread, bodies verbatim); registry wires gateway+service; market_analysis_server.py now a thin adapter (3 one-line tools, docstrings verbatim); stock_price_server.py cross-import repointed at the service (lines ~3179/3195/3212). fastmcp 3.2.3 confirmed: @mcp.tool() returns the original fn, so the old cross-calls were plain delegation — repoint is exact. Parity vs /tmp/micro_parity_before.json (AAPL): all fields identical incl. live-quote ones. ValueError contract preserved. 30 tests green vs test DB; 5 MCP servers import; Flask 38 routes |
 | 2 | Sentiment/News service | DONE | — | 2026-06-12 | quantcore/services/sentiment.py: NewsCollector moved wholesale from fastMCPTest/news_collector.py (RSS + yfinance fetch, FinBERT AutoModel loader) + SentimentService (collect_news/get_news_sentiment/get_sentiment_trend/list_news_symbols from news_sentiment_server; get_news from stock_price_server; get_security_news + get_sentiment_dashboard from REST routes). FinBERT duplicate KILLED: stock_price_server pipeline loader deleted; single loader in the service (scores byte-identical in parity check). YFinanceGateway.news() added. news_sentiment_server.py thin adapter (4 one-line tools); news_collector.py is a shim (options_analysis imports NewsCollector until Step 8); api/app.py /news + /news/sentiment-summary one call deep; stock_price_server internal news call repointed. Parity vs /tmp/sentiment_parity_before.json: only drift = snapshot captured_at (advances per call by design). 30 tests green; 5 servers import; Flask 38 routes |
 | 3 | Fundamentals service | DONE | — | 2026-06-12 | quantcore/services/fundamentals.py: FundamentalsService absorbs all 12 company_fundamentals tools + REST /earnings (get_earnings_dates) + watchlist report script (get_full_fundamental_profile). 8 YFinanceGateway methods added (info/financials/cashflow/quarterly_financials/quarterly_income_stmt/calendar/earnings_dates/history; info() is raw with no watchdog by design — preserves original t.info blocking semantics). FundamentalsRepository.ttl_seconds() added. Helpers refactored to take DataFrames instead of yf.Ticker (math verbatim; separate Ticker per property preserves per-call fetch behavior). Dead _qoq_vol_4 helper + redundant quarterly-revenue fetch in score compute dropped (one fewer network call, identical output). Pre-existing PROD BUG fixed separately (08b725f): cache_get_all_latest used SQLite-style GROUP BY rejected by Postgres — get_top_fundamental_stocks/get_upcoming_earnings/get_sector_fundamental_breakdown/get_fundamental_score_changes silently returned empty since the Postgres migration. company_fundamentals_server.py thin adapter (12 one-line tools, docstrings verbatim); stock_price_server cross-imports repointed (4 sites); api/app.py imports get_services at module level (replaces Step 2 local imports). Parity OK: 14 surfaces byte-identical + fresh-compute check (NVDA, TTL=0). 30 tests green; 5 servers import; Flask 37 routes |
-| 4 | Prices/technicals service + analytics | PENDING | — | — | |
+| 4 | Prices/technicals service + analytics | DONE | — | 2026-06-13 | quantcore/services/prices.py: PricesService absorbs 12 price/technical MCP tools (get_stock_price, get_rsi, get_macd, get_stochastic, get_volume_analysis, get_obv, get_vwap, get_candlestick_patterns, get_higher_lows, get_gap_analysis, get_historical_drawdown, get_vwap_history) + 4 REST surfaces (get_ohlcv_bars, get_technicals_table, get_technical_signals, screen_securities). quantcore/analytics/indicators.py NEW: pure safe_float/rsi_series/macd_series — single home for the REST-table RSI/MACD math (kills api/app.py's third copy). NOTE: MCP get_rsi/get_macd use adjust=False/different fetch periods and stay verbatim inside the service — intentionally NOT unified with the REST min_periods versions. OhlcvRepository.daily_bars_for_symbols() added for the screener's one-shot SQL. screen_securities takes filters dict + pre-loaded portfolio/watchlist lists (service stays free of Flask helpers; route passes _load_portfolio()/_load_watchlist()). registry hoists OptionsStore() to a shared local + wires prices=PricesService(ohlcv/yfinance/options/sentiment). stock_price_server.py: 12 tools now thin adapters (docstrings verbatim); _safe_int/_options_store/_chain_side_full KEPT (still used by options tools, migrate Step 5). api/app.py: 4 routes repointed; local _safe_float/_compute_rsi/_compute_macd helpers deleted. Parity 19/19 (only diff = non-deterministic set-iteration order in a ValueError string, identical content, present in original). 30 tests green; 5 MCP servers boot (stock-price 23 tools); Flask 37 routes; registry constructs PricesService |
 | 5 | Options service + PolygonGateway | PENDING | — | — | |
 | 6 | Harvester + Portfolio (DB positions, multi-owner) | PENDING | — | — | |
 | 7 | Recommendations service | PENDING | — | — | |
